// src/types.ts
type DefaultType<T = undefined> = T extends undefined ?
    {
        value: string | number,
        label?: string | number,
        color?: string,
        hide?: boolean,
        id?: string
        extraData?: string,
        checked?: string,
    } : {
        value: string | number,
        label?: string | number,
        color?: string,
        hide?: boolean,
        id?: string
        extraData?: string,
        checked?: string,
        data: T
    };

export type RnPickerOption<T = undefined> = DefaultType<T>;

export interface RnPickerProps {
    label: string;
    showLabel: boolean;
    pickerStyles?: ViewStyle;
    pickerTextStyles?: TextStyle;
    selectedOption: string | number;
    data: RnPickerOption[];
    disable?: boolean,
    onSelect: (item: RnPickerOption & { index?: number }) => void;
    required?: boolean;
    triggerValidation?: boolean;
    placeHolder?: string;
    customOptionRender?: (item, index) => JSX.Element;
    showModal?: boolean;
    hideSearch?: boolean;
    arrowColor?: string;
    showNoneOption?: boolean;
    changeSearchText?: Function,
}